% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example0.R
\docType{data}
\name{example0}
\alias{example0}
\title{Example}
\format{
An object of class \code{list} of length 4.
}
\usage{
example0
}
\description{
These example contains a dataset, the label of the observations in the dataset, the distance matrix of the dataset using L2 distance, and the edge matrix generated by 5-MST.
}
\details{
\describe{
\item{\code{data}}{pooled dataset of two samples sampling from two different t-distributions.}
\item{\code{label}}{label of the observations. 'sample 1' denotes the observations in sample 1. 'sample 2' denotes the observations in sample 2.}
\item{\code{distance}}{the distance matrix of the pooled dataset using L2 distance.}
\item{\code{edge}}{edge matrix generated by 5-MST.}
}
}
\keyword{datasets}
