% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdis.R
\name{getdis}
\alias{getdis}
\title{Get distance matrix}
\usage{
getdis(y)
}
\arguments{
\item{y}{dataset of the pooled data}
}
\value{
A distance matrix based on the L2 distance.
}
\description{
This function returns the distance matrix using L2 distance.
}
\examples{
data(example0)
data = as.matrix(example0$data)     # pooled dataset
getdis(data)

}
