% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmst.R
\name{kmst}
\alias{kmst}
\title{construct k-mst}
\usage{
kmst(y = NULL, dis = NULL, k = 1)
}
\arguments{
\item{y}{data}

\item{dis}{distance matrix}

\item{k}{parameter in K-MST, with default 1}
}
\value{
An edge matrix representing a similarity graph. Each row represents an edge and records the indices of two ends of an edge in two columns.
}
\description{
construct k-mst
}
\keyword{internal}
