% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permu_gen.R
\name{permu_gen}
\alias{permu_gen}
\title{get the test statistic and p-value based on permutation using robust generalized edge-count test}
\usage{
permu_gen(R1_list, R2_list, R1_test, R2_test, n_per)
}
\arguments{
\item{R1_list}{list of permuted weighted within-sample edge-counts of sample 1}

\item{R2_list}{list of permuted weighted within-sample edge-counts of sample 2}

\item{R1_test}{weighted within-sample edge-counts of sample 1}

\item{R2_test}{weighted within-sample edge-counts of sample 2}

\item{n_per}{number of permutations}
}
\value{
The p-value based on permutation distribution using robust generalized graph-based test.
}
\description{
get the test statistic and p-value based on permutation using robust generalized edge-count test
}
\keyword{internal}
