% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weiArith.R
\name{weiArith}
\alias{weiArith}
\title{Weighted function}
\usage{
weiArith(a, b)
}
\arguments{
\item{a}{node degree of one end of an edge}

\item{b}{node degree of another end of an edge}
}
\value{
The weight uses the arithmetic average of the node degrees of an edge.
}
\description{
This weight function returns the inverse of the arithmetic average of the node degrees of an edge.
}
\examples{
# For an edge where one end has a node degree of 5
# another end has a node degree of 6
 weiArith(6, 5)

}
