% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weiGeo.R
\name{weiGeo}
\alias{weiGeo}
\title{Weighted function}
\usage{
weiGeo(a, b)
}
\arguments{
\item{a}{node degree of one end of an edge}

\item{b}{node degree of another end of an edge}
}
\value{
The weight uses the geometric average of the node degrees of an edge.
}
\description{
This weight function returns the inverse of the geometric average of the node degrees of an edge.
}
\examples{
# For an edge where one end has a node degree of 5
# another end has a node degree of 6
weiGeo(6, 5)

}
