% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_R1R2.R
\name{weighted_R1R2}
\alias{weighted_R1R2}
\title{get weighted within-sample edge-counts and between-sample edge-counts}
\usage{
weighted_R1R2(E, n1, wei)
}
\arguments{
\item{E}{an edge matrix representing a similarity graph. Each row represents an edge and records the indices of two ends of an edge in two columns. The indices of observations in sample 1 are from 1 to n1 and indices of observations in sample 2 are from 1+n1 to n1+n2.}

\item{n1}{number of observations in sample 1.}

\item{wei}{a vector of weights of each edge.}
}
\value{
\item{R1}{the weighted within-sample edge-count for sample 1.}
\item{R2}{the weighted within-sample edge-count for sample 2.}
\item{R}{the weighted between-sample edge-count.}
}
\description{
get weighted within-sample edge-counts and between-sample edge-counts
}
\keyword{internal}
